Add portability framework from rra-c-util
Introduce the portability framework along with all the replacement
functions that will probably be needed.  Pull in all the IPv6 support
and networking portability functions as well as the basic C portability.
For now, keep all the old probes for our hand-rolled versions of things
that will eventually be handled by this framework.