Add IPv6 support when used with socket activation
The socket created by systemd is bound to both IPv4 and IPv6 by
default, but the networking code assumed that all IP addresses
were IPv4.  Support IPv6 internally so that this works correctly,
although currently there isn't a way for lbcd to handle IPv6 when
binding its own sockets when run as a normal daemon.