Rewrite lbcdclient
lbcdclient has been rewritten.  The output format is now completely
different and more human-readable.  The timeout and port are
configurable on the command line, and the option parsing now supports
long options.  The command no longer takes a list of multiple servers
to query, and exits with an error message if it cannot reach a server
within the timeout.  lbcdclient supports IPv6 if the IO::Socket::INET6
Perl module is installed.