Refactor lbcd packet handling code
Use an internal struct to hold some things, such as the client
address and text version of their address, so that we're not
passing them around everywhere as separate parameters or constantly
recomputing them.  Use a vector for the requested services, since
it's easier to manage.  Be more paranoid about the packet format
and require that it have exactly the correct size.