Fix systemd unit errors, service type
Switch the lbcd systemd unit to Type=simple and add lbcd.socket to
Also so that the two are linked.  There's no need to use notify for
lbcd since the only resource it needs to set up is its socket, which
socket activation handles.  Also fix a typo in the WantedBy field.